Implement support for exporting animated textures (i.e. animated Image Texture, Environment Texture nodes)
The following export options are available:
Export Animated Textures - When checked, Image/Environment Textures that are set to an animated Image Sequence are exported, for each frame between the Start and End Frame. If not enabled, the texture for the active scene's current frame is exported as a static texture path.

Start Frame - The start frame to export from.

End Frame - The end frame to export from.

The exported values for each frame match the settings in Blender's Texture nodes as calculated by the Frames, Start Frame, and Offset parameters.  The paths that are exported match the paths that Blender would render from for each frame from the expot Start Frame, to the End Frame.