When exporting lights, CreateNormalizeAttr() and set it to true.
This is so that exported lights from Blender will match when rendered externally with hdBlackbird, with no edits.  

A normalize toggle was added to hdBlackbird:
https://github.com/tangent-opensource/hdBlackbird/pull/144

However, since they are normalized in Blender/Cycles, the resulting render with no edits to the lights in hdBlackbird will mismatch without exporting the normalize attribute.